Moved category to filter parameters
